requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-Cors==4.0.0
requests==2.31.0
mysqlclient==2.2.4
python-dotenv==1.0.1
mysql-connector-python==8.3.0 
